+++ src/main/java/ovh/angrysoft/homedaemon/devices/traits/DoubleOutlet.java
@@ -1,4 +1,0 @@
-package ovh.angrysoft.homedaemon.devices.traits;
-
-public interface DoubleOutlet extends DoubleSwitch {
-}