+++ src/main/java/ovh/angrysoft/homedaemon/devices/traits/DoubleSwitch.java
@@ -1,12 +1,19 @@
 package ovh.angrysoft.homedaemon.devices.traits;
 
 public interface DoubleSwitch extends Trait {
-    void left(Boolean state);
-    void right(Boolean state);
-    boolean isLeftOn();
-    boolean isRightOn();
-    void toggleLeft();
-    void toggleRight();
+    void one(Boolean state);
+
+    void two(Boolean state);
+
+    boolean isOneOn();
+
+    boolean isTwoOn();
+
+    void toggleOne();
+
+    void toggleTwo();
+
     void allOn();
+
     void allOff();
 }